going towards sqlalchemy, full flask implementation, administering modules online
