+++ orchestra/rest/templates/admin/index.html
@@ -1,0 +1,10 @@
+{% extends 'admin/master.html' %}
+
+{% block body %}
+    <a href="#">Add module</a>
+    {% if modules %}
+        {% include 'module/table.html' %}
+    {% else %}
+        <p>No modules installed</p>
+    {% endif %}
+{% endblock %}
