+++ orchestra/rest/templates/module/item.html
@@ -1,0 +1,9 @@
+<tr>
+	<td>
+		<a href="#">{{ mid }}</a>
+	</td>
+	<td>
+		<a href="#">{{ module.metadata.name }}</a>
+	</td>
+	<td>{{ module.metadata.install.executable }}</td>
+</tr>
