+++ orchestra/rest/templates/module/table.html
@@ -1,0 +1,18 @@
+<table>
+	<thead>
+		<tr>
+			<th colspan="3">Modules</th>
+		</tr>
+		<tr>
+			<th>#</th>
+			<th>Name</th>
+			<th>Executable</th>
+		</tr>
+	</thead>
+	<tbody>
+		{% for mid,module in modules %}
+		    {% include 'module/item.html' %}
+		{% endfor %}
+	</tbody>
+
+</table>
