+++ requirements.txt
@@ -1,3 +1,7 @@
 flask
 flask_restful
+flask_admin
+flask_basicauth
+flask_sqlalchemy
+
 docker
